add arrays tests for numpy types
